+++ src/app/features/contact/contact-page.ts
@@ -1,7 +1,9 @@
 import { ChangeDetectionStrategy, Component } from '@angular/core';
+import { Container } from '../../shared/ui/container/container';
 
 @Component({
   selector: 'app-contact-page',
+  imports: [Container],
   templateUrl: './contact-page.html',
   styleUrl: './contact-page.scss',
   changeDetection: ChangeDetectionStrategy.OnPush,