+++ src/app/features/projects/case-studies/ridetogether-page.ts
@@ -1,7 +1,9 @@
 import { ChangeDetectionStrategy, Component } from '@angular/core';
+import { Container } from '../../../shared/ui/container/container';
 
 @Component({
   selector: 'app-ridetogether-page',
+  imports: [Container],
   templateUrl: './ridetogether-page.html',
   styleUrl: './ridetogether-page.scss',
   changeDetection: ChangeDetectionStrategy.OnPush,