+++ src/app/features/projects/projects-page.ts
@@ -1,9 +1,10 @@
 import { ChangeDetectionStrategy, Component } from '@angular/core';
 import { RouterLink } from '@angular/router';
+import { Container } from '../../shared/ui/container/container';
 
 @Component({
   selector: 'app-projects-page',
-  imports: [RouterLink],
+  imports: [RouterLink, Container],
   templateUrl: './projects-page.html',
   styleUrl: './projects-page.scss',
   changeDetection: ChangeDetectionStrategy.OnPush,